+++ app/src/main/java/com/example/se2_group4_project/backend/callbacks/ClientCallbacks.java
@@ -1,4 +1,4 @@
-package com.example.se2_group4_project.callbacks;
+package com.example.se2_group4_project.backend.callbacks;
 
 public interface ClientCallbacks {
     void onMessageSend(String send);
