+++ app/src/main/java/com/example/se2_group4_project/backend/database/crudInterface/RoomateDAO.java
@@ -1,6 +1,6 @@
-package com.example.se2_group4_project.database.crudInterface;
+package com.example.se2_group4_project.backend.database.crudInterface;
 
-import com.example.se2_group4_project.database.entities.Roomate;
+import com.example.se2_group4_project.backend.database.entities.Roomate;
 
 import androidx.room.Dao;
 import androidx.room.Delete;
