+++ app/src/main/java/com/example/se2_group4_project/backend/database/entities/Furnishings.java
@@ -1,12 +1,13 @@
-package com.example.se2_group4_project.database.entities;
+package com.example.se2_group4_project.backend.database.entities;
 
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
 
-@Entity
-public class Roomate {
-    @ColumnInfo(name = "roomMateID")
+@Entity(tableName = "furnishings")
+public class Furnishings {
+
+    @ColumnInfo(name = "furnishingsID")
     @PrimaryKey(autoGenerate = true)
     int id;
     @ColumnInfo(name = "playerID")
@@ -15,26 +16,20 @@ public class Roomate {
     int abilityID;
     @ColumnInfo(name = "name")
     String name;
-    @ColumnInfo(name = "trouble")
-    int trouble;
-    @ColumnInfo(name = "awake")
-    int awake;
+    @ColumnInfo(name = "clean")
+    int clean;
     @ColumnInfo(name = "dicePlace")
     int dicePlace;
     @ColumnInfo(name = "diceNeeded")
     int diceNeeded;
-    @ColumnInfo(name = "diceCount")
-    int diceCount;
 
-    public Roomate(int playerID, int abilityID, String name, int trouble, int awake, int dicePlace, int diceNeeded, int diceCount) {
+    public Furnishings(int playerID, int abilityID, String name, int clean, int dicePlace, int diceNeeded) {
         this.playerID = playerID;
         this.abilityID = abilityID;
         this.name = name;
-        this.trouble = trouble;
-        this.awake = awake;
+        this.clean = clean;
         this.dicePlace = dicePlace;
         this.diceNeeded = diceNeeded;
-        this.diceCount = diceCount;
     }
 
     public int getId() {
@@ -53,12 +48,8 @@ public class Roomate {
         return name;
     }
 
-    public int getTrouble() {
-        return trouble;
-    }
-
-    public int getAwake() {
-        return awake;
+    public int getClean() {
+        return clean;
     }
 
     public int getDicePlace() {
@@ -69,12 +60,8 @@ public class Roomate {
         return diceNeeded;
     }
 
-    public int getDiceCount() {
-        return diceCount;
-    }
-
-    public void setAwake(int awake) {
-        this.awake = awake;
+    public void setClean(int clean) {
+        this.clean = clean;
     }
 
     public void setDicePlace(int dicePlace) {
@@ -83,9 +70,5 @@ public class Roomate {
 
     public void setDiceNeeded(int diceNeeded) {
         this.diceNeeded = diceNeeded;
-    }
-
-    public void setDiceCount(int diceCount) {
-        this.diceCount = diceCount;
     }
 }
