--- app/src/main/java/com/example/se2_group4_project/callbacks/DatabaseCallbacks.java
@@ -1,8 +1,0 @@
-package com.example.se2_group4_project.callbacks;
-
-import com.example.se2_group4_project.database.entities.Player;
-
-public interface DatabaseCallbacks {
-    public void addPLayer(Player player);
-    public void getPlayer();
-}
